Fix windows build/test (#211)
* fix windows build/test


* attempt to fix linux build


* fix goling

+++ pkg/networkservice/common/mechanisms/kernel/server.go
@@ -33,12 +33,12 @@ type mechanismsServer struct{}
 
 func (m mechanismsServer) Request(ctx context.Context, req *networkservice.NetworkServiceRequest) (*networkservice.Connection, error) {
 	conn := req.Connection
-	inode := uint64(0)
+	inode := uintptr(0)
 	inode, err := fs.GetInode("/proc/self/net/ns")
 	if err != nil {
 		return nil, err
 	}
-	conn.Mechanism.Parameters[common.NetNSInodeKey] = strconv.FormatUint(inode, 10)
+	conn.Mechanism.Parameters[common.NetNSInodeKey] = strconv.FormatUint(uint64(inode), 10)
 	return next.Server(ctx).Request(ctx, req)
 }
